+++ firmware/NSPanelManagerFirmware/lib/NSPanel/NSPanel.hpp
@@ -45,7 +45,7 @@ public:
   void setDimLevel(uint8_t dimLevel);
   void setSleep(bool sleep);
   void setComponentText(const char *componentId, const char *text);
-  void setComponentVal(const char *componentId, uint8_t value);
+  void setComponentVal(const char *componentId, int16_t value);
   void setTimerTimeout(const char *componentId, uint16_t timeout);
   void setComponentPic(const char *componentId, uint8_t value);
   void setComponentPic1(const char *componentId, uint8_t value);
